runtime: joystream-utilities: fix comments and rename
