+++ Cargo.toml
@@ -17,7 +17,7 @@ members = [
 	"runtime-modules/constitution",
 	"runtime-modules/staking-handler",
     "runtime-modules/blog",
-    "runtime-modules/utilities",
+    "runtime-modules/joystream-utility",
 	"node",
 	"utils/chain-spec-builder/"
 ]
