+++ runtime/Cargo.toml
@@ -79,7 +79,7 @@ content-directory = { package = 'pallet-content-directory', default-features = f
 pallet_constitution = { package = 'pallet-constitution', default-features = false, path = '../runtime-modules/constitution' }
 staking-handler = { package = 'pallet-staking-handler', default-features = false, path = '../runtime-modules/staking-handler'}
 blog = { package = 'pallet-blog', default-features = false, path = '../runtime-modules/blog'}
-utilities = { package = 'pallet-utilities', default-features = false, path = '../runtime-modules/utilities'}
+joystream-utility = { package = 'joystream-utility', default-features = false, path = '../runtime-modules/joystream-utility'}
 
 [dev-dependencies]
 sp-io = { package = 'sp-io', default-features = false, git = 'https://github.com/paritytech/substrate.git', rev = 'a200cdb93c6af5763b9c7bf313fa708764ac88ca'}
@@ -154,7 +154,7 @@ std = [
     'pallet_constitution/std',
     'staking-handler/std',
     'blog/std',
-    'utilities/std'
+    'joystream-utility/std'
 ]
 runtime-benchmarks = [
     "frame-system/runtime-benchmarks",
@@ -171,7 +171,7 @@ runtime-benchmarks = [
     "proposals-discussion/runtime-benchmarks",
     "proposals-engine/runtime-benchmarks",
     "proposals-codex/runtime-benchmarks",
-    "utilities/runtime-benchmarks",
+    "joystream-utility/runtime-benchmarks",
     "pallet_constitution/runtime-benchmarks",
     "working-group/runtime-benchmarks",
     "forum/runtime-benchmarks",
