+++ runtime/src/integration/proposals/proposal_encoder.rs
@@ -34,7 +34,7 @@ impl ProposalEncoder<Runtime> for ExtrinsicProposalEncoder {
     fn encode_proposal(proposal_details: ProposalDetailsOf<Runtime>) -> Vec<u8> {
         let call = match proposal_details {
             ProposalDetails::Signal(signal) => {
-                Call::Utilities(utilities::Call::execute_signal_proposal(signal))
+                Call::JoystreamUtility(joystream_utility::Call::execute_signal_proposal(signal))
             }
             ProposalDetails::FundingRequest(params) => {
                 Call::Council(council::Call::funding_request(params))
@@ -42,9 +42,9 @@ impl ProposalEncoder<Runtime> for ExtrinsicProposalEncoder {
             ProposalDetails::SetMaxValidatorCount(new_validator_count) => Call::Staking(
                 pallet_staking::Call::set_validator_count(new_validator_count),
             ),
-            ProposalDetails::RuntimeUpgrade(blob) => {
-                Call::Utilities(utilities::Call::execute_runtime_upgrade_proposal(blob))
-            }
+            ProposalDetails::RuntimeUpgrade(blob) => Call::JoystreamUtility(
+                joystream_utility::Call::execute_runtime_upgrade_proposal(blob),
+            ),
             ProposalDetails::CreateWorkingGroupLeadOpening(create_opening_params) => {
                 wrap_working_group_call!(
                     create_opening_params.group,
@@ -58,7 +58,7 @@ impl ProposalEncoder<Runtime> for ExtrinsicProposalEncoder {
                 )
             }
             ProposalDetails::UpdateWorkingGroupBudget(amount, working_group, balance_kind) => {
-                Call::Utilities(utilities::Call::update_working_group_budget(
+                Call::JoystreamUtility(joystream_utility::Call::update_working_group_budget(
                     working_group,
                     amount,
                     balance_kind,
