+++ runtime/src/runtime_api.rs
@@ -295,7 +295,7 @@ impl_runtime_apis! {
             use crate::Council;
             use crate::Referendum;
             use crate::Blog;
-            use crate::Utilities;
+            use crate::JoystreamUtility;
 
 
             // Trying to add benchmarks directly to the Session Pallet caused cyclic dependency issues.
@@ -386,7 +386,7 @@ impl_runtime_apis! {
             add_benchmark!(params, batches, referendum, Referendum);
             add_benchmark!(params, batches, council, Council);
             add_benchmark!(params, batches, blog, Blog);
-            add_benchmark!(params, batches, utilities, Utilities);
+            add_benchmark!(params, batches, joystream_utility, JoystreamUtility);
 
             if batches.is_empty() { return Err("Benchmark not found for this pallet.".into()) }
             Ok(batches)
