+++ runtime/src/weights/mod.rs
@@ -27,11 +27,11 @@ pub mod pallet_utility;
 pub mod blog;
 pub mod council;
 pub mod forum;
+pub mod joystream_utility;
 pub mod membership;
 pub mod pallet_constitution;
 pub mod proposals_codex;
 pub mod proposals_discussion;
 pub mod proposals_engine;
 pub mod referendum;
-pub mod utilities;
 pub mod working_group;
